back to int

+++ tasks/lukin_i_elem_vec_sum/mpi/include/ops_mpi.hpp
@@ -18,7 +18,7 @@ class LukinIElemVecSumMPI : public BaseTask {
   bool RunImpl() override;
   bool PostProcessingImpl() override;
 
-  size_t vec_size_ = 0;
+  int vec_size_ = 0;
 };
 
 }  // namespace lukin_i_elem_vec_sum
